refactor: :construction: Adjusted design structure, organizing the imports and code format.
