+++ src/app.module.ts
@@ -9,20 +9,20 @@ import {
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { configService } from './infra/config/config.service';
 import { RolesMiddleware } from './infra/auth/roles.middleware';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 import { ThrottlerModule } from '@nestjs/throttler';
 import { ServeStaticModule } from '@nestjs/serve-static';
 import { join } from 'path';
 import { UserModule } from './modules/user/user.module';
+import { typeormConfig } from './infra/database/typeorm/typeorm.config';
 
 @Module({
     imports: [
         ServeStaticModule.forRoot({
             rootPath: join(__dirname, '..', 'public'),
         }),
-        TypeOrmModule.forRoot(configService.getTypeOrmConfig()),
+        TypeOrmModule.forRoot(typeormConfig.getTypeOrmConfig()),
         CacheModule.register(),
         ThrottlerModule.forRoot({
             ttl: 60 * 60,
